fix(24): Move errors to general errors file
