+++ src/app/api/users/route.ts
@@ -1,5 +1,5 @@
 import { UserController } from '@/lib/controllers/user.controller';
-import { ValidationError } from '@/lib/schemas/user.schema';
+import { ValidationError } from '@/lib/schemas/errors';
 import { type NextRequest, NextResponse } from 'next/server';
 
 const userController = new UserController();
